+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de s
 ### Adicionado
 - `QasGridGenerator`: Adicionado prop `useEllipsis` que por default é `true`, que irá adicionar a classe `ellipsis` em seu conteúdo.
 - `QasGridGenerator`: Adicionado prop `useInline` para mudar a disposição dos campos para ser por linha, ou seja, header e content ocupando a linha toda.
+- [`QasFormGenerator`, `QasGridGenerator`]: Adicionado prop `useCommonColumns` onde será possível passar um único objeto com seus breakpoints e será replicado para todos fields.
 
 ### Corrigido
 - `QasGridGenerator`: Corrigido ao passar `col` na prop columns(`{ lg: 'col' }`).